Begin journey towards sets of instructions
I think stamping is better off being implemented as a datastructure which
is then used to drive calls to PDFBox. This commit introduces some new
protocols which are meant to support this:

- CursorMovement
- Styling

The idea behind CursorMovement is that it can be used both in the selection
algorithm (think `horizontal-increase?`) and the conversion of tokens to
stamping instructions, because it provides an answer to the question of "do
I need to move the cursor explicitly based on this tokens width or height?"

Styling is usable both for implementing some of the other protocols, and for
providing an explicit view of which styling instructions need to be generated
for a token.

See the `Word` token for an initial implementation.