Update font handling according to the rules of PDFBox 2
The change in string length is such that a useful error is produced in case the
selected font for a piece of text does not contain glyphs for all necessary
characters in the string.

Because PDFBox now properly supports unicode, we can use the proper unicode
bullet character as the default for bullet lists.